requirements.txt
# Core
pyyaml>=6.0
pydantic>=2.5.0

# Voice
sounddevice>=0.4.6
numpy>=1.24
faster-whisper>=1.0.0
openwakeword>=0.6.0
silero-vad>=5.0
piper-tts>=1.2.0

# Brain / LLM
ollama>=0.3.0
httpx>=0.27.0

# Executor
pyautogui>=0.9.54
pygetwindow>=0.0.9
psutil>=5.9.0        # system_info (CPU, RAM, disco)
pycaw>=20240210      # system_volume (controle de áudio via COM)
comtypes>=1.2.0      # requerido pelo pycaw

# Utils
aiosqlite>=0.19.0

# Dev
pytest>=8.0.0
pytest-asyncio>=0.23.0
